Fix so that form data will be faked up to set any product in the database with a non-zero quantity but which isn't being displayed will have its quantity set to zero.
This deals with a member setting one or more ordered products to a
zero quantity, then previewing their order. At that point, the
zero-quantity product still has a non-zero value in the datbase (they
haven't saved the order yet), but isn't being displayed anymore, since
they set the quantity to zero and thus there is no form data to say
the quantity should be set to zero.



git-svn-id: http://subversion.assembla.com/svn/vokomokum/trunk@74 b3c4d6f4-6639-42c7-84c2-d72431877607